Adjusted names and locations of Conda envs
Depending on the system (if there is a shared file system on
a cluster) there can be clashes in the Conda env location
and Conda env naming. A single Conda environment could be used,
but there can also be issues with using the default /pw/.miniconda3
link to /.miniconda3. Here, explicitly insist on two different
Conda envs for the workflow host and the remote workers for
bypassing these issues.